+++ src/RefDocGen/MemberData/MethodData.cs
@@ -1,7 +1,6 @@
 using RefDocGen.DocExtraction;
-using RefDocGen.MemberData.Interfaces;
+using RefDocGen.MemberData.Abstract;
 using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Xml.Linq;
 
 namespace RefDocGen.MemberData;
@@ -9,69 +8,24 @@ namespace RefDocGen.MemberData;
 /// <summary>
 /// Represents data of a method.
 /// </summary>
-public record MethodData : ICallableMemberData
+/// <param name="MethodInfo"><see cref="System.Reflection.MethodInfo"/> object representing the method.</param>
+public record MethodData(MethodInfo MethodInfo) : InvokableMemberData(MethodInfo)
 {
-    /// <summary>
-    /// Create new <see cref="MethodData"/> instance.
-    /// </summary>
-    /// <param name="methodInfo"><see cref="System.Reflection.MethodInfo"/> object representing the field.</param>
-    public MethodData(MethodInfo methodInfo)
-    {
-        MethodInfo = methodInfo;
-        Parameters = MethodInfo.GetParameters().OrderBy(p => p.Position).Select(p => new MethodParameterData(p)).ToArray();
-    }
-
-    /// <summary>
-    /// <see cref="System.Reflection.MethodInfo"/> object representing the method.
-    /// </summary>
-    public MethodInfo MethodInfo { get; }
-
     /// <inheritdoc/>
-    public string Name => MethodInfo.Name;
+    public override string Name => MethodInfo.Name;
 
     /// <summary>
     /// Return type of the method.
     /// </summary>
     public string ReturnType => MethodInfo.ReturnType.Name;
 
-    /// <inheritdoc/>
-    public AccessModifier AccessModifier => AccessModifierExtensions.GetAccessModifier(MethodInfo.IsPrivate, MethodInfo.IsFamily,
-        MethodInfo.IsAssembly, MethodInfo.IsPublic, MethodInfo.IsFamilyAndAssembly, MethodInfo.IsFamilyOrAssembly);
-
-    /// <inheritdoc/>
-    public bool IsStatic => MethodInfo.IsStatic;
-
-    /// <inheritdoc/>
-    public bool IsOverridable => MethodInfo.IsVirtual && !MethodInfo.IsFinal;
-
-    /// <inheritdoc/>
-    public bool OverridesAnotherMember => !MethodInfo.Equals(MethodInfo.GetBaseDefinition());
-
-    /// <inheritdoc/>
-    public bool IsAbstract => MethodInfo.IsAbstract;
-
-    /// <inheritdoc/>
-    public bool IsFinal => MethodInfo.IsFinal;
-
-    /// <inheritdoc/>
-    public bool IsAsync => MethodInfo.GetCustomAttribute(typeof(AsyncStateMachineAttribute)) != null;
-
-    /// <inheritdoc/>
-    public bool IsSealed => OverridesAnotherMember && IsFinal;
-
-    /// <inheritdoc/>
-    public bool IsVirtual => MethodInfo.IsVirtual;
-
     /// <summary>
-    /// XML doc comment for the method
+    /// Documentation comment for the method return value.
     /// </summary>
-    public XElement DocComment { get; init; } = DocCommentTools.EmptySummaryNode;
+    public XElement ReturnValueDocComment { get; init; } = DocCommentTools.EmptyReturnsNode;
 
-    public XElement ReturnsDocComment { get; init; } = DocCommentTools.EmptyReturnsNode;
+    /// <inheritdoc/>
+    public override bool OverridesAnotherMember => !MethodInfo.Equals(MethodInfo.GetBaseDefinition());
 
-    /// <summary>
-    /// Gets the method parameters represented as <see cref="MethodParameterData"/> objects, ordered by their position.
-    /// </summary>
-    public MethodParameterData[] Parameters { get; }
 
 }
