+++ src/RefDocGen/TemplateGenerators/Default/Tools/Extensions/MethodParameterDataExtensions.cs
@@ -4,7 +4,7 @@ namespace RefDocGen.TemplateGenerators.Default.Tools.Extensions;
 
 internal static class MethodParameterDataExtensions
 {
-    internal static bool HasRefKeyword(this MethodParameterData parameterData)
+    internal static bool HasRefKeyword(this ParameterData parameterData)
     {
         return parameterData.IsPassedByReference && !parameterData.IsInput && !parameterData.IsOutput;
     }
